refactor: move test data to file
